[#50] feat : UserRegisterActivity에서 MainActivity로 이동
